Split oracle definitions into separate files.
